+++ CHANGELOG.md
@@ -9,6 +9,12 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!--## [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Add `alloydbomni` BETA resource and datasource
+- Add `aiven_alloydbomni_user` BETA resource and datasource
+- Add `aiven_alloydbomni_database` BETA resource and datasource
+
 ## [4.30.0] - 2024-12-05
 
 - Add `aiven_kafka_native_acl` resource